# Backend/identity_service/indentity/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: IdentityService API
  version: 1.0.0
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
paths:
...
  /vouchers/{code}:
    get:
      summary: Kiem tra voucher
      parameters:
        - in: path
          name: code
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
  /vouchers/{id}/use:
    post:
      summary: Ghi nhan voucher usage
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Success
...